ttskit/requirements.txt
scipy==1.3.1
xmltodict==0.12.0
phkit==0.2.7
#torch==1.7.0
numba==0.48.0
tqdm==4.36.1
visdom==0.1.8.9
multiprocess==0.70.11.1
webrtcvad==2.0.10
librosa==0.8.0
Unidecode==1.1.1
sounddevice==0.4.1
aukit==1.4.4
cycler==0.10.0
music21==6.1.0
tensorflow==1.15.2
pandas==0.25.1
inflect==4.1.0
matplotlib==3.1.1
setproctitle==1.2.2
numpy==1.16.5
umap_learn==0.1.1
pydub==0.24.1
dotmap==1.3.23
PyYaml==5.4.1
